refactor(Navigation): make it a component
